refactor: use sandbox processor for worker
