+++ deno.json
@@ -1,7 +1,7 @@
 {
   "tasks": {
-    "start": "deno run --allow-net --allow-env --allow-read --allow-ffi --unstable-sloppy-imports --allow-sys src/deno.ts",
-    "dev": "deno run --allow-net --allow-env --allow-read --allow-ffi --watch --unstable-sloppy-imports --allow-sys src/deno.ts"
+    "start": "deno run --allow-net --allow-env --allow-read --allow-ffi --unstable-sloppy-imports --allow-sys --allow-run src/deno.ts",
+    "dev": "deno run --allow-net --allow-env --allow-read --allow-ffi --watch --unstable-sloppy-imports --allow-sys --allow-run src/deno.ts"
   },
   "imports": {
     "hono": "npm:hono@4.7.2",
